394
Adding Authorization Checks

+++ controllers/feed.js
@@ -127,6 +127,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -154,10 +159,15 @@ exports.deletePost = (req, res, next) => {
         const error = new Error('Could nor find post.');
         error.statusCode = 404;
         throw error;
-        // Check logged in user
-        clearImage(post.imageUrl);
-        return Post.findByIdAndRemove(postId);
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
+      // Check logged in user
+      clearImage(post.imageUrl);
+      return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
       console.log(result);
